// frontend/src/api/generated/model/webPublicationManifest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * crossbill API
 * OpenAPI spec version: 0.0.0
 */
import type { ReadiumLink } from './readiumLink.ts';
import type { ReadiumMetadata } from './readiumMetadata.ts';

/**
 * A complete Readium Web Publication Manifest.
 */
export interface WebPublicationManifest {
  '@context'?: string;
  metadata: ReadiumMetadata;
  links: ReadiumLink[];
  readingOrder: ReadiumLink[];
  resources: ReadiumLink[];
  toc: ReadiumLink[];
}